'update-CRUD-product'

+++ MapogoSport-BE/src/main/java/mapogo/entity/ProductDetail.java
@@ -31,7 +31,7 @@ public class ProductDetail implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ProductDetailId")
-	private Long productDetailId;
+	private Integer productDetailId;
 
 	@ManyToOne
 	@JoinColumn(name = "ProductId", nullable = false)
